Add files via upload
Refer to the information given here and answer questions relating to the Florist Problem.						
A florist has 420 acres of land on which she plants roses and sunflowers.						
For each acre of rose planted, her expenses are $40 and for each acre of sunflower planted, her expenses are $120.						
Each acre of rose requires 100 units of storage and yields a profit of $70. 						
Each acre of sunflower requires 45 units of storage and yields a profit of $95. 						
If the total amount of storage space available is 19,500 units and the florist can spend a maximum of $30,000,						
 how many acres of each flower should she plant in order to maximize her profit?  Partial acres are not allowed, that is, acres should be in integers.